# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Health Insurance Risk Classifier API
  version: 0.1.0
  description: |
    Provides endpoints for service health, single risk prediction, and model metadata.
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /v1/evaluations/risk:
    post:
      tags: [evaluations]
      summary: Evaluate insurance risk for an individual
      operationId: createRiskEvaluation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskEvaluationRequest'
            example:
              age: 42
              sex: male
              bmi: 31.2
              children: 2
              smoker: 'no'
      responses:
        '200':
          description: Risk evaluation generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskEvaluationResponse'
              example:
                evaluation_id: 9e3f50f8-66d1-4bf2-b5c8-33f8d0319c6d
                risk_category: Medium
                model_version: rule-v1
                created_at: '2026-03-21T12:00:00Z'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: bad_request
                message: children must be between 0 and 10
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: validation_error
                message: bmi is required
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: internal_error
                message: unexpected error while generating risk evaluation
...
components:
...
  schemas:
...
    RiskEvaluationRequest:
      type: object
      required: [age, sex, bmi, children, smoker]
      additionalProperties: false
      properties:
        age:
          type: integer
          minimum: 18
          maximum: 120
        sex:
          $ref: '#/components/schemas/Sex'
        bmi:
          type: number
          format: float
        children:
          type: integer
          minimum: 0
          maximum: 10
        smoker:
          $ref: '#/components/schemas/SmokerStatus'
    RiskEvaluationResponse:
      type: object
      required: [evaluation_id, risk_category, model_version, created_at]
      properties:
        evaluation_id:
          type: string
          format: uuid
        risk_category:
          $ref: '#/components/schemas/RiskCategory'
        model_version:
          type: string
          example: rule-v1
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
...
    RiskCategory:
      type: string
      enum: [Low, Medium, High]
    Sex:
      type: string
      enum: [male, female]
    SmokerStatus:
      type: string
      enum: ['yes', 'no']